# frontend/openapi.json
{
  "components" : {
    "schemas" : {
...
      "CurrencySummary" : {
        "properties" : {
          "byGameType" : {
            "description" : "The game types with finished games, in catalog order",
            "items" : {
              "$ref" : "#/components/schemas/GameTypeSummary"
            },
            "type" : "array"
          },
          "currencyCode" : {
            "type" : "string"
          },
          "inPlay" : {
            "$ref" : "#/components/schemas/InPlay"
          },
          "total" : {
            "$ref" : "#/components/schemas/StatsFigures",
            "description" : "Every finished game in this currency"
          }
        },
        "required" : [ "byGameType", "currencyCode", "inPlay", "total" ],
        "type" : "object"
      },
      "FinishGameRequest" : {
        "properties" : {
          "bounty" : {
            "description" : "Bounties won. Not for cash games. Defaults to 0",
            "minimum" : 0,
            "type" : [ "number", "null" ]
          },
          "prize" : {
            "description" : "Cash won, bounties apart. Cash game: amount when leaving the table. Defaults to 0",
            "minimum" : 0,
            "type" : [ "number", "null" ]
          },
          "ticketDescription" : {
            "maxLength" : 150,
            "minLength" : 0,
            "type" : [ "string", "null" ]
          },
          "ticketPrizeValue" : {
            "description" : "Value of a tournament ticket won. Not for cash games. Defaults to 0",
            "minimum" : 0,
            "type" : [ "number", "null" ]
          }
        },
        "type" : "object"
      },
      "GameRequest" : {
        "properties" : {
          "bounty" : {
            "description" : "Bounties won. Defaults to 0",
            "minimum" : 0,
            "type" : [ "number", "null" ]
          },
          "buyIn" : {
            "description" : "Price of one entry, also when it was paid with a ticket. Cash game: amount brought to the table",
            "minimum" : 0,
            "type" : "number"
          },
          "entries" : {
            "description" : "Entries including re-entries. Defaults to 1",
            "format" : "int32",
            "minimum" : 1,
            "type" : [ "integer", "null" ]
          },
          "gameType" : {
            "enum" : [ "TOURNAMENT", "SIT_AND_GO", "CASH" ],
            "type" : "string"
          },
          "modality" : {
            "description" : "Defaults to NLHE",
            "enum" : [ "NLHE", "PLO" ],
            "type" : [ "string", "null" ]
          },
          "name" : {
            "maxLength" : 150,
            "minLength" : 0,
            "type" : [ "string", "null" ]
          },
          "notes" : {
            "maxLength" : 5000,
            "minLength" : 0,
            "type" : [ "string", "null" ]
          },
          "paidWithTicket" : {
            "description" : "One entry was paid with a ticket instead of cash. Defaults to false",
            "type" : [ "boolean", "null" ]
          },
          "playedAt" : {
            "description" : "Optional local start time (HH:mm), to order the games of a day",
            "example" : "21:30",
            "type" : [ "string", "null" ]
          },
          "playedOn" : {
            "format" : "date",
            "type" : "string"
          },
          "prize" : {
            "description" : "Cash won, bounties apart. Cash game: amount when leaving the table. Defaults to 0",
            "minimum" : 0,
            "type" : [ "number", "null" ]
          },
          "roomId" : {
            "format" : "int64",
            "type" : "integer"
          },
          "status" : {
            "description" : "IN_PLAY (registered when it starts, no result yet) or FINISHED. When omitted: FINISHED if a result is sent (prize, bounty or ticket); otherwise IN_PLAY for a new game and unchanged for an existing one",
            "enum" : [ "IN_PLAY", "FINISHED" ],
            "type" : [ "string", "null" ]
          },
          "ticketDescription" : {
            "maxLength" : 150,
            "minLength" : 0,
            "type" : [ "string", "null" ]
          },
          "ticketPrizeValue" : {
            "description" : "Value of a tournament ticket won as a prize; informative, not part of net. Defaults to 0",
            "minimum" : 0,
            "type" : [ "number", "null" ]
          },
          "variantId" : {
            "description" : "Optional; must be a variant of the game type",
            "format" : "int64",
            "type" : [ "integer", "null" ]
          }
        },
        "required" : [ "buyIn", "gameType", "playedOn", "roomId" ],
        "type" : "object"
      },
      "GameResponse" : {
        "properties" : {
          "bounty" : {
            "type" : "number"
          },
          "buyIn" : {
            "type" : "number"
          },
          "createdAt" : {
            "format" : "date-time",
            "type" : "string"
          },
          "currencyCode" : {
            "description" : "Currency of every amount of this game (the one of its room)",
            "type" : "string"
          },
          "entries" : {
            "format" : "int32",
            "type" : "integer"
          },
          "gameType" : {
            "enum" : [ "TOURNAMENT", "SIT_AND_GO", "CASH" ],
            "type" : "string"
          },
          "id" : {
            "format" : "int64",
            "type" : "integer"
          },
          "invested" : {
            "description" : "Money paid for the entries (an entry paid with a ticket costs nothing)",
            "type" : "number"
          },
          "modality" : {
            "enum" : [ "NLHE", "PLO" ],
            "type" : "string"
          },
          "name" : {
            "type" : [ "string", "null" ]
          },
          "net" : {
            "description" : "Real money won or lost: prize + bounty - invested",
            "type" : "number"
          },
          "notes" : {
            "type" : [ "string", "null" ]
          },
          "paidWithTicket" : {
            "type" : "boolean"
          },
          "playedAt" : {
            "example" : "21:30:00",
            "type" : [ "string", "null" ]
          },
          "playedOn" : {
            "format" : "date",
            "type" : "string"
          },
          "prize" : {
            "type" : "number"
          },
          "room" : {
            "$ref" : "#/components/schemas/RoomRef"
          },
          "status" : {
            "description" : "IN_PLAY while the result is not known yet; its buy-in already counts in net",
            "enum" : [ "IN_PLAY", "FINISHED" ],
            "type" : "string"
          },
          "ticketDescription" : {
            "type" : [ "string", "null" ]
          },
          "ticketPrizeValue" : {
            "type" : "number"
          },
          "updatedAt" : {
            "format" : "date-time",
            "type" : "string"
          },
          "variant" : {
            "anyOf" : [ {
              "$ref" : "#/components/schemas/VariantRef"
            }, {
              "type" : "null"
            } ]
          }
        },
        "required" : [ "bounty", "buyIn", "createdAt", "currencyCode", "entries", "gameType", "id", "invested", "modality", "net", "paidWithTicket", "playedOn", "prize", "room", "status", "ticketPrizeValue", "updatedAt" ],
        "type" : "object"
      },
      "GameTypeSummary" : {
        "properties" : {
          "figures" : {
            "$ref" : "#/components/schemas/StatsFigures"
          },
          "gameType" : {
            "enum" : [ "TOURNAMENT", "SIT_AND_GO", "CASH" ],
            "type" : "string"
          }
        },
        "required" : [ "figures", "gameType" ],
        "type" : "object"
      },
...
      "PageResponseGameResponse" : {
        "properties" : {
          "items" : {
            "items" : {
              "$ref" : "#/components/schemas/GameResponse"
            },
            "type" : "array"
          },
          "page" : {
            "description" : "Zero-based page number",
            "format" : "int32",
            "type" : "integer"
          },
          "size" : {
            "description" : "Requested page size",
            "format" : "int32",
            "type" : "integer"
          },
          "totalItems" : {
            "format" : "int64",
            "type" : "integer"
          },
          "totalPages" : {
            "format" : "int32",
            "type" : "integer"
          }
        },
        "required" : [ "items", "page", "size", "totalItems", "totalPages" ],
        "type" : "object"
      },
...
      "RebuyRequest" : {
        "properties" : {
          "amount" : {
            "description" : "Amount added to the buy-in of the game",
            "exclusiveMinimum" : 0,
            "type" : "number"
          }
        },
        "required" : [ "amount" ],
        "type" : "object"
      },
...
      "RoomRef" : {
        "properties" : {
          "id" : {
            "format" : "int64",
            "type" : "integer"
          },
          "name" : {
            "type" : "string"
          }
        },
        "required" : [ "id", "name" ],
        "type" : "object"
      },
...
      "StatsSummaryResponse" : {
        "properties" : {
          "currencies" : {
            "items" : {
              "$ref" : "#/components/schemas/CurrencySummary"
            },
            "type" : "array"
          }
        },
        "required" : [ "currencies" ],
        "type" : "object"
      },
      "VariantCreateRequest" : {
        "properties" : {
          "gameType" : {
            "enum" : [ "TOURNAMENT", "SIT_AND_GO", "CASH" ],
            "type" : "string"
          },
          "name" : {
            "maxLength" : 100,
            "minLength" : 0,
            "type" : "string"
          }
        },
        "required" : [ "gameType", "name" ],
        "type" : "object"
      },
      "VariantRef" : {
        "properties" : {
          "code" : {
            "type" : [ "string", "null" ]
          },
          "id" : {
            "format" : "int64",
            "type" : "integer"
          },
          "name" : {
            "type" : [ "string", "null" ]
          }
        },
        "required" : [ "id" ],
        "type" : "object"
      },
      "VariantResponse" : {
        "properties" : {
          "active" : {
            "type" : "boolean"
          },
          "builtIn" : {
            "type" : "boolean"
          },
          "code" : {
            "description" : "Code of a built-in variant, translated by the client; null for user-defined ones",
            "type" : [ "string", "null" ]
          },
          "gameType" : {
            "enum" : [ "TOURNAMENT", "SIT_AND_GO", "CASH" ],
            "type" : "string"
          },
          "id" : {
            "format" : "int64",
            "type" : "integer"
          },
          "inUse" : {
            "description" : "The variant is used by games: it cannot be deleted",
            "type" : "boolean"
          },
          "name" : {
            "description" : "Name of a user-defined variant; null for built-in ones",
            "type" : [ "string", "null" ]
          }
        },
        "required" : [ "active", "builtIn", "gameType", "id", "inUse" ],
        "type" : "object"
      },
      "VariantUpdateRequest" : {
        "properties" : {
          "active" : {
            "description" : "Inactive variants keep their history but are not offered for new games",
            "type" : "boolean"
          },
          "name" : {
            "description" : "New name of a user-defined variant. Must be omitted (or null) for built-in variants.",
            "maxLength" : 100,
            "minLength" : 0,
            "pattern" : ".*\\S.*",
            "type" : [ "string", "null" ]
          }
        },
        "required" : [ "active" ],
        "type" : "object"
      }
    }
  },
  "info" : {
    "description" : "REST API of poker-bankroll, a self-hosted poker bankroll manager.",
    "license" : {
      "name" : "MIT",
      "url" : "https://opensource.org/licenses/MIT"
    },
    "title" : "poker-bankroll API"
  },
  "openapi" : "3.1.0",
  "paths" : {
...
    "/games" : {
      "get" : {
        "description" : "All filters are optional and combined with AND.",
        "operationId" : "list_2",
        "parameters" : [ {
          "description" : "Played on or after this date",
          "in" : "query",
          "name" : "from",
          "required" : false,
          "schema" : {
            "format" : "date",
            "type" : "string"
          }
        }, {
          "description" : "Played on or before this date",
          "in" : "query",
          "name" : "to",
          "required" : false,
          "schema" : {
            "format" : "date",
            "type" : "string"
          }
        }, {
          "in" : "query",
          "name" : "gameType",
          "required" : false,
          "schema" : {
            "enum" : [ "TOURNAMENT", "SIT_AND_GO", "CASH" ],
            "type" : "string"
          }
        }, {
          "in" : "query",
          "name" : "modality",
          "required" : false,
          "schema" : {
            "enum" : [ "NLHE", "PLO" ],
            "type" : "string"
          }
        }, {
          "in" : "query",
          "name" : "roomId",
          "required" : false,
          "schema" : {
            "format" : "int64",
            "type" : "integer"
          }
        }, {
          "in" : "query",
          "name" : "variantId",
          "required" : false,
          "schema" : {
            "format" : "int64",
            "type" : "integer"
          }
        }, {
          "in" : "query",
          "name" : "status",
          "required" : false,
          "schema" : {
            "enum" : [ "IN_PLAY", "FINISHED" ],
            "type" : "string"
          }
        }, {
          "description" : "Currency of the room, e.g. EUR",
          "in" : "query",
          "name" : "currency",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        }, {
          "description" : "Text contained in the name or the notes, ignoring case",
          "in" : "query",
          "name" : "q",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        }, {
          "description" : "Zero-based page number",
          "in" : "query",
          "name" : "page",
          "required" : false,
          "schema" : {
            "default" : 0,
            "format" : "int32",
            "minimum" : 0,
            "type" : "integer"
          }
        }, {
          "in" : "query",
          "name" : "size",
          "required" : false,
          "schema" : {
            "default" : 50,
            "format" : "int32",
            "maximum" : 200,
            "minimum" : 1,
            "type" : "integer"
          }
        }, {
          "description" : "`<field>,<asc|desc>` with field one of playedOn, net, buyIn, prize, createdAt",
          "in" : "query",
          "name" : "sort",
          "required" : false,
          "schema" : {
            "default" : "playedOn,desc",
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PageResponseGameResponse"
                }
              }
            },
            "description" : "OK"
          }
        },
        "summary" : "List games",
        "tags" : [ "Games" ]
      },
      "post" : {
        "description" : "With only the required fields the game is recorded as in play; send a result (or `status: FINISHED`) to record it finished.",
        "operationId" : "create_2",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/GameRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "201" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/GameResponse"
                }
              }
            },
            "description" : "Created"
          }
        },
        "summary" : "Record a game",
        "tags" : [ "Games" ]
      }
    },
    "/games/{id}" : {
      "delete" : {
        "operationId" : "delete_2",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "format" : "int64",
            "type" : "integer"
          }
        } ],
        "responses" : {
          "204" : {
            "description" : "No Content"
          }
        },
        "summary" : "Delete a game",
        "tags" : [ "Games" ]
      },
      "get" : {
        "operationId" : "get_1",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "format" : "int64",
            "type" : "integer"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/GameResponse"
                }
              }
            },
            "description" : "OK"
          }
        },
        "summary" : "Get a game",
        "tags" : [ "Games" ]
      },
      "put" : {
        "description" : "Replaces every field; omitted optional fields take their default.",
        "operationId" : "update_2",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "format" : "int64",
            "type" : "integer"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/GameRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/GameResponse"
                }
              }
            },
            "description" : "OK"
          }
        },
        "summary" : "Update a game",
        "tags" : [ "Games" ]
      }
    },
    "/games/{id}/finish" : {
      "post" : {
        "description" : "Sets its result; with no body (or an empty one) nothing was won.",
        "operationId" : "finish",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "format" : "int64",
            "type" : "integer"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/FinishGameRequest"
              }
            }
          }
        },
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/GameResponse"
                }
              }
            },
            "description" : "OK"
          }
        },
        "summary" : "Finish a game in play",
        "tags" : [ "Games" ]
      }
    },
    "/games/{id}/re-entries" : {
      "post" : {
        "description" : "Adds one entry, paid in cash.",
        "operationId" : "addReEntry",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "format" : "int64",
            "type" : "integer"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/GameResponse"
                }
              }
            },
            "description" : "OK"
          }
        },
        "summary" : "Add a re-entry to a tournament or Sit&Go in play",
        "tags" : [ "Games" ]
      }
    },
    "/games/{id}/rebuys" : {
      "post" : {
        "description" : "Adds the amount to the buy-in.",
        "operationId" : "addRebuy",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "format" : "int64",
            "type" : "integer"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/RebuyRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/GameResponse"
                }
              }
            },
            "description" : "OK"
          }
        },
        "summary" : "Add a rebuy to a cash game in play",
        "tags" : [ "Games" ]
      }
    },
...
    "/stats/summary" : {
      "get" : {
        "description" : "Takes the filters of the games list, all optional and combined with AND. Only finished games count in the figures; games in play are reported apart.",
        "operationId" : "summary",
        "parameters" : [ {
          "description" : "Played on or after this date",
          "in" : "query",
          "name" : "from",
          "required" : false,
          "schema" : {
            "format" : "date",
            "type" : "string"
          }
        }, {
          "description" : "Played on or before this date",
          "in" : "query",
          "name" : "to",
          "required" : false,
          "schema" : {
            "format" : "date",
            "type" : "string"
          }
        }, {
          "in" : "query",
          "name" : "gameType",
          "required" : false,
          "schema" : {
            "enum" : [ "TOURNAMENT", "SIT_AND_GO", "CASH" ],
            "type" : "string"
          }
        }, {
          "in" : "query",
          "name" : "modality",
          "required" : false,
          "schema" : {
            "enum" : [ "NLHE", "PLO" ],
            "type" : "string"
          }
        }, {
          "in" : "query",
          "name" : "roomId",
          "required" : false,
          "schema" : {
            "format" : "int64",
            "type" : "integer"
          }
        }, {
          "in" : "query",
          "name" : "variantId",
          "required" : false,
          "schema" : {
            "format" : "int64",
            "type" : "integer"
          }
        }, {
          "description" : "Currency of the room, e.g. EUR",
          "in" : "query",
          "name" : "currency",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        }, {
          "description" : "Text contained in the name or the notes, ignoring case",
          "in" : "query",
          "name" : "q",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/StatsSummaryResponse"
                }
              }
            },
            "description" : "OK"
          }
        },
        "summary" : "Results overall and per game type, for each currency",
        "tags" : [ "Statistics" ]
      }
    },
    "/variants" : {
      "get" : {
        "description" : "Ordered by game type, built-in variants first, then user-defined ones by name.",
        "operationId" : "list",
        "parameters" : [ {
          "in" : "query",
          "name" : "gameType",
          "required" : false,
          "schema" : {
            "enum" : [ "TOURNAMENT", "SIT_AND_GO", "CASH" ],
            "type" : "string"
          }
        }, {
          "in" : "query",
          "name" : "active",
          "required" : false,
          "schema" : {
            "type" : "boolean"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "items" : {
                    "$ref" : "#/components/schemas/VariantResponse"
                  },
                  "type" : "array"
                }
              }
            },
            "description" : "OK"
          }
        },
        "summary" : "List variants",
        "tags" : [ "Variants" ]
      },
      "post" : {
        "operationId" : "create",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/VariantCreateRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "201" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/VariantResponse"
                }
              }
            },
            "description" : "Created"
          }
        },
        "summary" : "Create a user-defined variant",
        "tags" : [ "Variants" ]
      }
    },
    "/variants/{id}" : {
      "delete" : {
        "description" : "Only when no game uses it; otherwise deactivate it.",
        "operationId" : "delete",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "format" : "int64",
            "type" : "integer"
          }
        } ],
        "responses" : {
          "204" : {
            "description" : "No Content"
          }
        },
        "summary" : "Delete a user-defined variant",
        "tags" : [ "Variants" ]
      },
      "put" : {
        "description" : "Built-in variants can only be activated or deactivated; user-defined ones can also be renamed.",
        "operationId" : "update",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "format" : "int64",
            "type" : "integer"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/VariantUpdateRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/VariantResponse"
                }
              }
            },
            "description" : "OK"
          }
        },
        "summary" : "Update a variant",
        "tags" : [ "Variants" ]
      }
    }
  },
...
}